# docs/swagger.yaml
basePath: /
definitions:
  model.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
...
  req.RequestAddSchedule:
    properties:
      atTime:
        type: string
      day:
        type: string
      every:
        type: string
      jobName:
        type: string
    required:
    - atTime
    - day
    - every
    - jobName
    type: object
...
  req.RequestDeleteSchedule:
    properties:
      scheduleId:
        type: string
    required:
    - scheduleId
    type: object
  req.RequestDeleteTelegram:
    properties:
      telegramName:
        type: string
    required:
    - telegramName
    type: object
  req.RequestInfoTelegram:
    properties:
      chatId:
        type: string
      disableNotification:
        type: string
      telegramName:
        type: string
      telegramToken:
        type: string
    required:
    - chatId
    - disableNotification
    - telegramName
    - telegramToken
    type: object
...
host: localhost:7000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Tool Testing API
  version: "1.0"
paths:
  /alert/add-telegram:
    post:
      consumes:
      - application/json
      parameters:
      - description: alert
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RequestInfoTelegram'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Add thông tin telegram để alert
      tags:
      - Alert-service
  /alert/delete-telegram:
    post:
      consumes:
      - application/json
      parameters:
      - description: alert
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RequestDeleteTelegram'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: xóa thông tin telegram để alert
      tags:
      - Alert-service
  /alert/list-telegram:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.Response'
      summary: Hiển thị thông tin telegram để alert
      tags:
      - Alert-service
...
  /schedule/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: schedule
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RequestAddSchedule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Add schedule để check testing
      tags:
      - schedule-service
  /schedule/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: schedule
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RequestDeleteSchedule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Xóa schedule để check testing
      tags:
      - schedule-service
  /schedule/list:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Hiện thị các schedule
      tags:
      - schedule-service
...
swagger: "2.0"